gpdbrestore --truncate will not try to truncate a nonexistent table.
Previously, if the set of tables to be restored using gpdbrestore --truncate
included a table that did not exist in the target database, gpdbrestore would
raise an exception when attempting to truncate the nonexistent table.  Now,
gpdbrestore displays a warning regarding the nonexistent table, and the table is
restored normally.

Authors: Jimmy Yih and James McAtamney

+++ gpMgmt/bin/gppylib/operations/restore.py
@@ -415,7 +415,10 @@ def truncate_restore_tables(restore_tables, master_port, dbname):
                 qry = 'Truncate %s' % t
                 execSQL(conn, qry)
             except Exception as e:
-                raise Exception("Could not truncate table %s.%s: %s" % (dbname, t, str(e).replace('\n', '')))
+                if 'relation "%s" does not exist' % t in str(e).replace('\n', ''):
+                    logger.warning("Skipping truncate of %s.%s because the relation does not exist." % (dbname, t))
+                else:
+                    raise Exception("Could not truncate table %s.%s: %s" % (dbname, t, str(e).replace('\n', '')))
 
         conn.commit()
     except Exception as e:
